Install needed CMake modules
Recuse build into Modules dir and install CMake modules and tools
that clients of Canvas need to build things like dictionaries.

Note that FindTBB is also installed to help re-find this dependency
if needed for imported targets. Long term, it should be promoted to
cetbuildtools2 (or similar), or only used internally by canvasConfig
to avoid clashes with other possible instances.